# Use a lower optimization level